requirements.txt
streamlit==1.22.0
pandas==1.5.3
pymongo==4.3.3
certifi==2023.5.7
python-dotenv==1.0.0
dnspython==2.3.0